docker環境の構築と不具合箇所の修正
