+++ backend/app/src/templates/index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Hello from Python</title>
+        <style>
+            #firstMsg{
+                font-size: 30px;
+                font-family:Impact;
+                color:blue
+            }
+            #secondMsg{
+                font-size: 30px;
+                font-family:Impact;
+                color:yellow
+            }
+        </style>
+    </head>
+    <body bgcolor="black">
+        <p id="firstMsg">
+            {{msg}}
+        </p>
+        <p id="secondMsg">
+            {{additional_msg}}
+        </p>
+    </body>
+</html>
