+++ backend/docker-compose.yml
@@ -1,0 +1,7 @@
+version: "3"
+services:
+  pythonapp:
+    container_name: "app"
+    build: ./app
+    ports:
+      - "8080:8080"
